Raspberry_Pi/requirements.txt
pyserial
flask
fastapi[standard]

